+++ include/circle/soundbasedevice.h
@@ -2,7 +2,7 @@
 // soundbasedevice.h
 //
 // Circle - A C++ bare metal environment for Raspberry Pi
-// Copyright (C) 2017  R. Stange <rsta2@o2online.de>
+// Copyright (C) 2017-2018  R. Stange <rsta2@o2online.de>
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -92,7 +92,7 @@ public:
 	/// \return Number of bytes consumed
 	/// \note Not used, if GetChunk() is overloaded.
 	/// \note Can be called on any core.
-	int Write (const void *pBuffer, unsigned nCount);
+	int Write (const void *pBuffer, size_t nCount);
 
 	/// \return Queue size in number of frames
 	/// \note Not used, if GetChunk() is overloaded.
